implemented get/set tests as in #473 #474 #475 #476 #480 #481 #482 #483

+++ trunk/goograde/goograde.sql
@@ -59,10 +59,10 @@ CREATE TABLE Assignments (
 -- entity table for GradingRules class
 CREATE TABLE GradingRules (    
 		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		aFloor FLOAT,
-		bFloor FLOAT,
-		cFloor FLOAT,
-		dFloor FLOAT
+		aFloor INTEGER,
+		bFloor INTEGER,
+		cFloor INTEGER,
+		dFloor INTEGER
 		);
